[merge] :: push changes form App.vue to Vqb.vue and delete App.vue
