+++ tests/unit/vqb.spec.ts
@@ -10,38 +10,44 @@ localVue.use(Vuex);
 
 describe('Vqb', () => {
   it('should instantiate', () => {
-    const wrapper = shallowMount(Vqb, { store: setupStore(), localVue });
+    const store = setupStore();
+    const wrapper = shallowMount(Vqb, { store, localVue });
     expect(wrapper.exists()).toBeTruthy();
-    expect(wrapper.vm.$data.isEditingStep).toBeFalsy();
+    expect(store.state.isEditingStep).toBeFalsy();
   });
 
   it('should set editingMode on when step is created', () => {
     const store = setupStore();
     const wrapper = shallowMount(Vqb, { store, localVue });
-    expect(wrapper.vm.$data.isEditingStep).toBeFalsy();
+    expect(store.state.isEditingStep).toBeFalsy();
     wrapper.find(DataViewer).vm.$emit('stepCreated');
-    expect(wrapper.vm.$data.isEditingStep).toBeTruthy();
+    expect(store.state.isEditingStep).toBeTruthy();
   });
 
   it('should set pipeline when form is saved', () => {
-    const store = setupStore();
+    const store = setupStore({
+      pipeline: [{ name: 'domain', domain: 'foo' }],
+      isEditingStep: true,
+    });
     const wrapper = shallowMount(Vqb, { store, localVue });
-    wrapper.setData({ isEditingStep: true });
     wrapper
       .find(FormRenameStep)
       .vm.$emit('formSaved', { name: 'rename', oldname: 'columnA', newname: 'columnAA' });
-    expect(wrapper.vm.$data.isEditingStep).toBeFalsy();
+    expect(store.state.isEditingStep).toBeFalsy();
     expect(store.state.pipeline).toEqual([
+      { name: 'domain', domain: 'foo' },
       { name: 'rename', oldname: 'columnA', newname: 'columnAA' },
     ]);
   });
 
   it('should cancel edition', () => {
-    const store = setupStore({ pipeline: [{ name: 'domain', domain: 'foo' }] });
+    const store = setupStore({
+      pipeline: [{ name: 'domain', domain: 'foo' }],
+      isEditingStep: true,
+    });
     const wrapper = shallowMount(Vqb, { store, localVue });
-    wrapper.setData({ isEditingStep: true });
     wrapper.find(FormRenameStep).vm.$emit('cancel');
-    expect(wrapper.vm.$data.isEditingStep).toBeFalsy();
+    expect(store.state.isEditingStep).toBeFalsy();
     expect(store.state.pipeline).toEqual([{ name: 'domain', domain: 'foo' }]);
   });
 });
